Ember.js by Oren Rubin
Ember brings a paradigm shift when it comes to MVC frameworks. Its
Apple origin gave it all the advantages of a robust (native like)
framework dedicated for large scale applications, while recent changes
added sugar making it easier to learn.
This lecture is not just for new Embers, but also for those familiar
with her, and wish to have a deeper understanding, ramping up their
skills with best practices.